test(split-button): typescript refactor
